# docs/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Expoflores API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/me/exhibitor": {
      "get": {
        "tags": [
          "representante"
        ],
        "summary": "My Exhibitor",
        "operationId": "my_exhibitor_api_v1_me_exhibitor_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExhibitorDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/me/quota": {
      "get": {
        "tags": [
          "representante"
        ],
        "summary": "My Quota",
        "description": "Cupo total vs. usado, desglosado por categoria. Calculado, nunca leido de una columna.",
        "operationId": "my_quota_api_v1_me_quota_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyQuota"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/me/participants": {
      "get": {
        "tags": [
          "representante"
        ],
        "summary": "List Participants",
        "operationId": "list_participants_api_v1_me_participants_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "Exhibitor",
                    "Guest",
                    "Service"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ParticipantRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "representante"
        ],
        "summary": "Create Participant",
        "operationId": "create_participant_api_v1_me_participants_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParticipantIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipantRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/participants/{participant_id}": {
      "get": {
        "tags": [
          "representante"
        ],
        "summary": "Get Participant",
        "operationId": "get_participant_api_v1_me_participants__participant_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "participant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Participant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipantRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "representante"
        ],
        "summary": "Update Participant",
        "operationId": "update_participant_api_v1_me_participants__participant_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "participant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Participant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParticipantUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipantRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "representante"
        ],
        "summary": "Delete Participant",
        "operationId": "delete_participant_api_v1_me_participants__participant_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "participant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Participant Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/participants": {
      "get": {
        "tags": [
          "participantes"
        ],
        "summary": "List Participants",
        "operationId": "list_participants_api_v1_participants_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "exhibitor_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exhibitor Id"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "Exhibitor",
                    "Guest",
                    "Service"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ParticipantWithExhibitor_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/rules/stand-sizes": {
      "get": {
        "tags": [
          "reglas"
        ],
        "summary": "Stand Sizes",
        "operationId": "stand_sizes_api_v1_rules_stand_sizes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StandSizeRuleRead"
                  },
                  "type": "array",
                  "title": "Response Stand Sizes Api V1 Rules Stand Sizes Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/rules/credentials": {
      "get": {
        "tags": [
          "reglas"
        ],
        "summary": "Credentials",
        "operationId": "credentials_api_v1_rules_credentials_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CredentialRuleRead"
                  },
                  "type": "array",
                  "title": "Response Credentials Api V1 Rules Credentials Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "ContactRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "email": {
            "type": "string",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "phone",
          "email"
        ],
        "title": "ContactRead"
      },
      "CredentialRuleRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "credentials_per_block": {
            "type": "integer",
            "title": "Credentials Per Block"
          },
          "block_m2": {
            "type": "integer",
            "title": "Block M2"
          },
          "rounding_mode": {
            "type": "string",
            "title": "Rounding Mode"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category",
          "credentials_per_block",
          "block_m2",
          "rounding_mode"
        ],
        "title": "CredentialRuleRead"
      },
...
      "ExhibitorDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "tax_id": {
            "type": "string",
            "title": "Tax Id"
          },
          "tax_id_type": {
            "type": "string",
            "title": "Tax Id Type"
          },
          "legal_name": {
            "type": "string",
            "title": "Legal Name"
          },
          "stand_name": {
            "type": "string",
            "title": "Stand Name"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "requested_m2": {
            "type": "integer",
            "title": "Requested M2"
          },
          "stand_category": {
            "type": "string",
            "title": "Stand Category"
          },
          "quota": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Quota"
          },
          "assigned": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Assigned"
          },
          "available": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Available"
          },
          "representative": {
            "$ref": "#/components/schemas/RepresentativeRead"
          },
          "contacts": {
            "items": {
              "$ref": "#/components/schemas/ContactRead"
            },
            "type": "array",
            "title": "Contacts"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tax_id",
          "tax_id_type",
          "legal_name",
          "stand_name",
          "address",
          "requested_m2",
          "stand_category",
          "quota",
          "assigned",
          "available",
          "representative",
          "contacts"
        ],
        "title": "ExhibitorDetail"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MyQuota": {
        "properties": {
          "exhibitor_id": {
            "type": "integer",
            "title": "Exhibitor Id"
          },
          "legal_name": {
            "type": "string",
            "title": "Legal Name"
          },
          "stand_name": {
            "type": "string",
            "title": "Stand Name"
          },
          "requested_m2": {
            "type": "integer",
            "title": "Requested M2"
          },
          "stand_category": {
            "type": "string",
            "title": "Stand Category"
          },
          "quota": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Quota"
          },
          "assigned": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Assigned"
          },
          "available": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Available"
          },
          "participants_total": {
            "type": "integer",
            "title": "Participants Total"
          },
          "participants_without_email": {
            "type": "integer",
            "title": "Participants Without Email"
          }
        },
        "type": "object",
        "required": [
          "exhibitor_id",
          "legal_name",
          "stand_name",
          "requested_m2",
          "stand_category",
          "quota",
          "assigned",
          "available",
          "participants_total",
          "participants_without_email"
        ],
        "title": "MyQuota",
        "description": "Cupo del stand propio, desglosado por categoria (§4)."
      },
...
      "Page_ParticipantRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ParticipantRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[ParticipantRead]"
      },
      "Page_ParticipantWithExhibitor_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ParticipantWithExhibitor"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[ParticipantWithExhibitor]"
      },
      "ParticipantIn": {
        "properties": {
          "first_name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Last Name"
          },
          "identification": {
            "type": "string",
            "maxLength": 20,
            "minLength": 1,
            "title": "Identification"
          },
          "identification_type": {
            "type": "string",
            "enum": [
              "CEDULA",
              "RUC",
              "PASSPORT",
              "FOREIGN_ID"
            ],
            "title": "Identification Type"
          },
          "phone": {
            "type": "string",
            "maxLength": 30,
            "minLength": 1,
            "title": "Phone"
          },
          "position": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Position"
          },
          "category": {
            "type": "string",
            "enum": [
              "Exhibitor",
              "Guest",
              "Service"
            ],
            "title": "Category"
          },
          "provider_company": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Company"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "first_name",
          "last_name",
          "identification",
          "identification_type",
          "phone",
          "position",
          "category"
        ],
        "title": "ParticipantIn"
      },
      "ParticipantRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "identification": {
            "type": "string",
            "title": "Identification"
          },
          "identification_type": {
            "type": "string",
            "title": "Identification Type"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "position": {
            "type": "string",
            "title": "Position"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "provider_company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Company"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "credential_notified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credential Notified At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "first_name",
          "last_name",
          "identification",
          "identification_type",
          "phone",
          "position",
          "category",
          "provider_company",
          "email"
        ],
        "title": "ParticipantRead"
      },
      "ParticipantUpdate": {
        "properties": {
          "first_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "identification": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Identification"
          },
          "identification_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "CEDULA",
                  "RUC",
                  "PASSPORT",
                  "FOREIGN_ID"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Identification Type"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 30,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "position": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "category": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "Exhibitor",
                  "Guest",
                  "Service"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "provider_company": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Company"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          }
        },
        "type": "object",
        "title": "ParticipantUpdate",
        "description": "PATCH: el invariante de `provider_company` se revalida en el servicio sobre el\nresultado de la mezcla, porque aqui no se conoce la categoria que ya tiene la fila."
      },
...
      "RepresentativeRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "identification": {
            "type": "string",
            "title": "Identification"
          },
          "identification_type": {
            "type": "string",
            "title": "Identification Type"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "position": {
            "type": "string",
            "title": "Position"
          }
        },
        "type": "object",
        "required": [
          "id",
          "full_name",
          "identification",
          "identification_type",
          "email",
          "phone",
          "position"
        ],
        "title": "RepresentativeRead"
      },
...
      "StandSizeRuleRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "min_m2": {
            "type": "integer",
            "title": "Min M2"
          },
          "max_m2": {
            "type": "integer",
            "title": "Max M2"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "min_m2",
          "max_m2"
        ],
        "title": "StandSizeRuleRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}